Move the book builder code into book/ folder out of docs/
