+++ book/lib/fast-array-onebuf.mjs
