+++ book/lib/fast-dict-onebuf.mjs
