+++ book/lib/fast-indirect-objects.mjs
