+++ book/lib/fast-number-to-string.mjs
