+++ book/lib/fast-parse-number.mjs
