+++ book/lib/fast-parse-object.mjs
