+++ book/lib/fast-pdfnumber-pool.mjs
