+++ book/lib/fast-refs-class.mjs
