+++ book/lib/fast-refs.mjs
