+++ book/lib/fast-size-in-bytes.mjs
