+++ book/lib/fast-sync-load.mjs
