+++ book/lib/paged.browser.js
