+++ book/lib/progress-handler.js
